+++ Automate.Model/src/Movables/MovableCapabilities.cs
@@ -3,8 +3,8 @@
     public class MovableCapabilities
     {
         public int MaxCarryWeight { get; set; } = 10;
-        public int MovementSpeed { get; set; } = 1;
-        public int WorkSpeed { get; set; } = 1;
+        public float MovementSpeed { get; set; } = 1;
+        public float WorkSpeed { get; set; } = 1;
         public int Intelligence { get; set; } = 0;
     }
 }
